Fact Analyzer
This project work has shown the power of using simulation for understanding a manufacturing system. We have used the Fact Analyzer tool for creating the flow of the manufacturing system based on the data provided. We understood that batch size and batching affect the whole system. And how the whole system is dependent on the smooth flow of products from one operation to another. Also observed is how buffers and stores play a role in the smooth flow of the products when there is a stagnant flow of the manufacturing system.